requirements.txt
langchain>=0.1.6
langchain-google-genai>=0.0.5
langchain-community>=0.0.15
langchain-core>=0.1.17
langchain-tavily>=0.1.2
python-dotenv>=1.0.0
google-generativeai>=0.3.1
tavily-python>=0.3.0
wikipedia>=1.4.0
arxiv>=2.0.0
